requirements.txt
absl-py==2.3.1
accelerate==1.11.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albumentations==1.3.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.11.0
attrs==25.4.0
basicsr==1.4.2
certifi==2025.10.5
charset-normalizer==3.4.4
clean-fid==0.1.35
click==8.3.0
clip-anytorch==2.6.0
compel==2.0.1
contourpy==1.3.3
controlnet-aux==0.0.6
cycler==0.12.1
diffusers==0.35.2
einops==0.3.0
facexlib==0.3.0
fastapi==0.121.0
filelock==3.20.0
filterpy==1.4.5
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
gfpgan==1.3.8
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
invisible-watermark==0.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonmerge==1.9.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
k-diffusion==0.0.12
kiwisolver==1.4.9
kornia==0.6.0
lazy_loader==0.4
llvmlite==0.45.1
lmdb==1.7.5
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.1.1
open-clip-torch==2.0.2
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.3
peft==0.17.1
picklescan==0.0.31
piexif==1.1.3
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pydantic==2.12.3
pydantic_core==2.41.4
pyDeprecate==0.3.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
qudida==0.0.4
realesrgan==0.3.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
resize-right==0.0.2
rpds-py==0.28.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tb-nightly==2.21.0a20251023
tensorboard==2.20.0
tensorboard-data-server==0.7.2
test_tube==0.7.5
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.21
tokenizers==0.22.1
torch==2.9.0
torchdiffeq==0.2.5
torchmetrics==0.6.0
torchruntime==1.24.0
torchsde==0.2.6
torchvision==0.24.0
tqdm==4.67.1
trampoline==0.1.2
transformers==4.57.1
triton==3.5.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wandb==0.22.3
wcwidth==0.2.14
Werkzeug==3.1.3
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
